[Update]: 소셜 회원가입 및 로그인 API Request 변경 (#90)
socialId 추가
socialType 삭제

Related to: #89
