+++ .gitignore
@@ -181,4 +181,4 @@ application-dev.yml
 application-prod.yml
 
 # firebase admin sdk
-sobok-76d0a-firebase-adminsdk-qb2ez-cedea5e056.json
+soboksobok-564b3-firebase-adminsdk-l4k48-58b981ed69.json
