+++ src/main/java/io/sobok/SobokSobok/auth/ui/dto/SocialLoginRequest.java
@@ -1,19 +1,13 @@
 package io.sobok.SobokSobok.auth.ui.dto;
 
-import io.sobok.SobokSobok.auth.domain.SocialType;
 import io.swagger.v3.oas.annotations.media.Schema;
-import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
 
 @Builder
 public record SocialLoginRequest(
 
         @Schema(requiredMode = Schema.RequiredMode.REQUIRED)
-        String code,
-
-        @Schema(requiredMode = Schema.RequiredMode.REQUIRED)
-        @NotNull
-        SocialType socialType,
+        String socialId,
 
         @Schema(requiredMode = Schema.RequiredMode.REQUIRED)
         String deviceToken
